requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
scikit-learn==1.2.2
pandas==2.0.0
prometheus-client==0.16.0
joblib==1.2.0
pydantic==1.10.7
httpx
kafka-python
